--- viewer/src/components/Snackbar.tsx
@@ -1,32 +1,0 @@
-import { type SnackbarKey, SnackbarProvider, closeSnackbar, enqueueSnackbar } from "notistack";
-import React from "react";
-
-export function InfoSnackbar(props: { message: string }) {
-  const hideSnackbar = (snackbarId: SnackbarKey) => (
-    <>
-      <button
-        type={"button"}
-        onClick={() => {
-          closeSnackbar(snackbarId);
-        }}
-      >
-        Dismiss
-      </button>
-    </>
-  );
-
-  React.useEffect(() => {
-    enqueueSnackbar(props.message, { action: hideSnackbar });
-  });
-
-  return (
-    <div>
-      <SnackbarProvider
-        anchorOrigin={{ horizontal: "right", vertical: "top" }}
-        autoHideDuration={null}
-        variant={"warning"}
-        preventDuplicate={true}
-      />
-    </div>
-  );
-}
